+++ engine/flink/avro-util/src/main/scala/pl/touk/nussknacker/engine/avro/kryo/AvroSerializersRegistrar.scala
@@ -2,9 +2,11 @@ package pl.touk.nussknacker.engine.avro.kryo
 
 import com.typesafe.config.Config
 import com.typesafe.scalalogging.LazyLogging
+import net.ceedubs.ficus.Ficus.{toFicusConfig, _}
 import org.apache.avro.generic.GenericData
 import org.apache.flink.api.common.ExecutionConfig
 import org.apache.flink.api.java.typeutils.AvroUtils
+import pl.touk.nussknacker.engine.api.component.ComponentProviderConfig
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.client.{CachedConfluentSchemaRegistryClientFactory, ConfluentSchemaRegistryClientFactory}
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.kryo.SchemaIdBasedAvroGenericRecordSerializer
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.serialization.GenericRecordSchemaIdSerializationSupport
@@ -16,27 +18,42 @@ class AvroSerializersRegistrar extends SerializersRegistrar with LazyLogging {
 
   override def register(modelConfig: Config, executionConfig: ExecutionConfig): Unit = {
     logger.debug("Registering default avro serializers")
+    val resolvedKafkaConfig = resolveConfig(modelConfig)
     AvroUtils.getAvroUtils.addAvroSerializersIfRequired(executionConfig, classOf[GenericData.Record])
-    registerGenericRecordSchemaIdSerializationForGlobalKafkaConfigIfNeed(modelConfig, executionConfig)
+    registerGenericRecordSchemaIdSerializationForGlobalKafkaConfigIfNeed(resolvedKafkaConfig, executionConfig)
   }
 
-  // It registers GenericRecordSchemaIdSerialization only for global kafka config (in 'kafka' path)
+  private def resolveConfig(modelConfig: Config): Option[KafkaConfig] = {
+    val componentsConfig = modelConfig.getAs[Map[String, ComponentProviderConfig]]("components").getOrElse(Map.empty)
+    val componentsKafkaConfigs = componentsConfig
+      .filterNot(_._2.disabled)
+      .map(_._2.config)
+      .flatMap(KafkaConfig.parseConfigOpt(_, "config.kafka"))
+    val theOnlyOneEnabledKafkaConfigOpt = componentsKafkaConfigs match {
+      case firstEnabledKafkaConfig :: Nil => Some(firstEnabledKafkaConfig)
+      case _ => None // mechanism would be disabled in case if there is more than one kafka component enabled
+    }
+    theOnlyOneEnabledKafkaConfigOpt.orElse(KafkaConfig.parseConfigOpt(modelConfig))
+  }
+
+  // It registers GenericRecordSchemaIdSerialization only for first kafka component config
   // If you want to register GenericRecordSchemaIdSerialization for other kafka config you need to invoke
   // `AvroSerializersRegistrar.registerGenericRecordSchemaIdSerializationIfNeed` directly
-  private def registerGenericRecordSchemaIdSerializationForGlobalKafkaConfigIfNeed(modelConfig: Config, executionConfig: ExecutionConfig): Unit = {
-    KafkaConfig.parseConfigOpt(modelConfig).foreach { globalKafkaConfig =>
-      val autoRegister = globalKafkaConfig.kafkaEspProperties
+  private def registerGenericRecordSchemaIdSerializationForGlobalKafkaConfigIfNeed(componentKafkaConfig: Option[KafkaConfig], executionConfig: ExecutionConfig): Unit = {
+    componentKafkaConfig.foreach { kafkaConfig =>
+      val autoRegister = kafkaConfig.kafkaEspProperties
         .flatMap(_.get(AvroSerializersRegistrar.autoRegisterRecordSchemaIdSerializationProperty).map(_.toBoolean))
         .getOrElse(true)
+      logger.debug(s"Auto registering SchemaIdBasedAvroGenericRecordSerializer: $autoRegister")
       if (autoRegister) {
-        AvroSerializersRegistrar.registerGenericRecordSchemaIdSerializationIfNeed(executionConfig, globalKafkaConfig)
+        AvroSerializersRegistrar.registerGenericRecordSchemaIdSerializationIfNeed(executionConfig, kafkaConfig)
       }
     }
   }
 
 }
 
-object AvroSerializersRegistrar extends LazyLogging{
+object AvroSerializersRegistrar extends LazyLogging {
 
   // This property is mainly for testing purpose. You can use it to skip GenericRecordSchemaIdSerialization registration
   val autoRegisterRecordSchemaIdSerializationProperty = "autoRegisterRecordSchemaIdSerialization"