+++ engine/flink/avro-util/src/test/scala/pl/touk/nussknacker/engine/avro/source/flink/TestFromFileSpec.scala
@@ -37,6 +37,7 @@ class TestFromFileSpec extends FunSuite with Matchers with LazyLogging {
   private lazy val config = ConfigFactory.empty()
     .withValue("kafka.kafkaAddress", fromAnyRef("notused:1111"))
     .withValue("kafka.kafkaProperties.\"schema.registry.url\"", fromAnyRef("notused:2222"))
+    .withValue("kafka.avroKryoGenericRecordSchemaIdSerialization", fromAnyRef(false))
 
   test("Should pass correct timestamp from test data") {