+++ engine/flink/generic/src/test/scala/pl/touk/nussknacker/genericmodel/GenericClassExtractionTest.scala
@@ -9,8 +9,8 @@ class GenericClassExtractionTest extends ClassExtractionBaseTest {
 
   protected override val model: LocalModelData = {
     val config = ConfigFactory.load()
-      .withValue("kafka.kafkaAddress", fromAnyRef("notused:1111"))
-      .withValue("kafka.kafkaProperties.\"schema.registry.url\"", fromAnyRef("notused:1111"))
+      .withValue("components.kafka.config.kafkaAddress", fromAnyRef("notused:1111"))
+      .withValue("components.kafka.config.kafkaProperties.\"schema.registry.url\"", fromAnyRef("notused:1111"))
     LocalModelData(config, new GenericConfigCreator)
   }
   protected override val outputResource = "/extractedTypes/genericCreator.json"